Moved the metrics back into each controller and added expose behaviour so only the running components metrics would be published
